Finishing up adding quiet param to gets
Incidentally takes care of issue #9 to replace garbage slices with
io.LimitReader and io.Copy to ioutil.Discard.

As well this commit has a couple bug fixes pertaining to the
interaction with memcached. When a miss occurred on a single chunk
(because of evictions or what have you) the program would lock up
because it wasn’t properly reading in the response.