Adding a NoopEnd parameter to the get request struct
This tracks if a Noop was the end of the pipelined gets or if a GET
was. If a Noop was, we have to respond with an extra NOOP header at the
end or the client can’t know we’re done.

Also this incidentally led to fixing a bug with the regular get
parsing, where the quiet param was empty and thus causing index out of
bounds panics.