First pass at adding a get-and-touch (GAT) command
Unfortunately, the abstraction breaks here where the memcached protocol
diverges. The text protocol has no get-and-touch command, while the
binary does. The interface supports the operation, but the text
protocol does not. In this case, the text parser should never return a
GATCommand so the text responder has no implementation.