First pass at get-and-touch in testing client.
Again the text protocol does not support the GAT command, so the
implementation is simply to panic. The abstraction is a bit leaky,
though, since the text protocol gets different treatment in the
blast.go script than the binary protocol. It’s worth it, though, given
the simplicity elsewhere. The machinery is the same, the inputs are
different.