Proper error handling for bad chunk tokens on get and GAT
This bug has proven to be difficult to track down. The changes here
keep track of errors, misses, or missed because of errors and returns
the proper response to the caller after guaranteeing that all of the
response has been ingested from memcached. Previously it was leaving
some data unread, so the next read would get bad metadata and think
there were billions of data chunks. Now all reads should be clean.