First pass at counter metrics
The implementation is simply a map of string to *uint64. sync/atomic is
used to alter values. The counters map is protected by a sync.RWMutex
to prevent concurrent alterations. The only Writer in this case (in the
RWMutex sense) is the function to create a new metric. Otherwise, map
accesses are all reads.