Properly pass opaque tokens
Fixes #17

Opaque tokens are meant to be returned on the response that corresponds
to the request that contained it. Previously all responses had an
opaque token of 0. Now opaque tokens are sent through any path from
request parsing to response *except* on error with gets. Get errors are
a bit wonky because they always end up causing a disconnect from
everyone anyway, so passing the opaque token has no value. In this case
if we end up being able to respond  without being in a bad state, the
response will have an opaque token of 0.