Adding SetQ, AddQ, and ReplaceQ support.
There is no optimization to read in multiple requests as there is in
GetQ for batch gets. Each request, if pipelined or not, is serviced and
responded to before the next is read. Again, mutations (including sets
and the like) are all in an async path for most clients so this is fine
for the Netflix use case.