Adding explicit support for a noop command
Noop is used when clients reconnect to verify the connection. This
happens frequently in the cloud, so this is necessary to implement.
Previously noop was only used as an endcap for a batch get (as an
alternative to a regular GET) but now it's a first-class citizen.

The text protocol bit is a small extension since the original protocol
does not support text noops.