Handlers have the responsibility of connecting now.
This allows the server to use handlers that do not need to create a
connection. The connections are still assumed to be unix domain
sockets. If remote connections are needed later, the ability can be
added later. The API is still a bit leaky for handlers that do not need
to connect, but those can simply ignore the parameter.