Calculating data chunk size based on memcached overhead and key size
The previous implementation took a fixed data size of 1024 bytes, which
would spill over to another slab if the key was too long. Now the
overhead of memcached metadata and key length are taken into account
when chunking data so there is no space wastage and everything stays in
the same slab.

In other words the data now fits *just* right.

Fixes #57