Delete socket path before binding. Also, allow rend to listen on unix sockets.
In order to test the socket binding, I had to add in options to listen
on a unix socket. I might as well leave them in as they might be useful
and don't get in the way.

to listen on a unix domain socket:

`--use-domain-socket --sock-path "/path/to/socket"`

fixes #64