Adding ability to share locks between different locked orchestrators
This means that operations on different servers running on different ports (IPs, what-have-you) will share a set of locks such that a set on one port blocks read access on the other. This guarantees consistency in a world where there aremultiple servers for multiple use cases. The implementation is also written such that there can be many (or few) of such lock sets to share across many different servers, if one was so inclined to do so.