Updating histograms to have less overhead when flipping buffers
This basically came down to keeping less state unnecessarily, only holding on to the old buffer. As well, there was no need to do a bunch of atomic stores when I could just make a new struct with the old buffer. This avoids unnecessary coordination overhead and allocations in the critical section.