Removing req.Quiet changes in the L1L2Batch orchestrator.
Under high load, rend would receive SETQ calls, which do not expect a response. When the client got a response, it assumed a problem and disconnected. This is exactly the worst thing to do under heavy load as it requires clients to reconnect, which takes even longer. This change removes the erroneous erasure of the quiet flag.