Add .travis.yml config for Travis CI support.
The default options for Go projects might be sufficient as per instructions.
This config simply adds the setting to use the most-recent Linux distro
available on Travis CI (Trusty) and disables `sudo` to use the more efficient
container-based infrastructure.

Additional customizations can be added once this config is enabled and it's easy
to test new iterations via the GitHub PR mechanism, which will get Travis to
test each PR automatically.